requirements.txt
pytesseract
pdf2image
azure-storage-blob
python-dotenv